finishing sebelum crsf token di order update

+++ app/api/category/delete/route.ts
@@ -2,13 +2,20 @@ import { NextResponse } from "next/server";
 import connectDB from "@/lib/dbConnect";
 import Category from "@/models/Category";
 import Product from "@/models/Product";
+import getCookieToken from "@/utils/getCookieToken";
 
 export async function POST(req: Request) {
   try {
     await connectDB();
+    const csrfTokenFromCookie = getCookieToken(req, "csrf_token");
 
     // Ambil data dari body request
-    const { id } = await req.json();
+    const { id, csrfToken } = await req.json();
+
+    if (!csrfTokenFromCookie || !csrfToken || csrfTokenFromCookie !== csrfToken) {
+      return NextResponse.json({ message: "Invalid CSRF" }, { status: 403 });
+    }
+
 
     // Mencari dan menghapus kategori berdasarkan ID
     const deletedCategory = await Category.findByIdAndDelete(id);
